+++ dessia_common/breakdown.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Code for breakdowns
-
+Code for breakdowns.
 """
 
 import sys