+++ dessia_common/core.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-dessia_common
-
+Module to handle serialization for engineering objects.
 """
 
 import time
@@ -45,6 +44,8 @@ import dessia_common.utils.helpers as dch
 import dessia_common.files as dcf
 
 _FORBIDDEN_ARGNAMES = ['self', 'cls', 'progress_callback', 'return']
+
+_fullargsspec_cache = {}
 
 
 class DessiaObject(SerializableObject):
@@ -324,8 +325,14 @@ class DessiaObject(SerializableObject):
         """
         Transform serialized argument of a method to python objects ready to use in method evaluation.
         """
-        method_object = getattr(self, method)
-        args_specs = inspect.getfullargspec(method_object)
+        method_full_name = f'{self.full_classname}.{method}'
+        if method_full_name in _fullargsspec_cache:
+            args_specs = _fullargsspec_cache[method_full_name]
+        else:
+            method_object = getattr(self, method)
+            args_specs = inspect.getfullargspec(method_object)
+            _fullargsspec_cache[method_full_name] = args_specs
+
         allowed_args = args_specs.args[1:]
 
         arguments = {}
@@ -415,7 +422,13 @@ class DessiaObject(SerializableObject):
 
     def __copy__(self):
         """ Generic copy use inits of objects. """
-        class_argspec = inspect.getfullargspec(self.__class__)
+        class_name = self.full_classname
+        if class_name in _fullargsspec_cache:
+            class_argspec = _fullargsspec_cache[class_name]
+        else:
+            class_argspec = inspect.getfullargspec(self.__class__)
+            _fullargsspec_cache[class_name] = class_argspec
+
         dict_ = {}
         for arg in class_argspec.args:
             if arg != 'self':
@@ -428,7 +441,13 @@ class DessiaObject(SerializableObject):
 
     def __deepcopy__(self, memo=None):
         """ Generic deep copy use inits of objects. """
-        class_argspec = inspect.getfullargspec(self.__class__)
+        class_name = self.full_classname
+        if class_name in _fullargsspec_cache:
+            class_argspec = _fullargsspec_cache[class_name]
+        else:
+            class_argspec = inspect.getfullargspec(self.__class__)
+            _fullargsspec_cache[class_name] = class_argspec
+
         if memo is None:
             memo = {}
         dict_ = {}
@@ -721,6 +740,7 @@ class PhysicalObject(DessiaObject):
 
 class MovingObject(PhysicalObject):
     """ A 3D object which display can move down a path from according to defined steps. """
+
     def volmdlr_primitives_step_frames(self):
         """ Return a list of volmdlr primitives to build up volume model. """
         raise NotImplementedError('Object inheriting MovingObject should implement volmdlr_primitives_step_frames')
@@ -734,6 +754,7 @@ class MovingObject(PhysicalObject):
 
 class Parameter(DessiaObject):
     """ A value from a Parameter Set. """
+
     def __init__(self, lower_bound, upper_bound, periodicity=None, name=''):
         DessiaObject.__init__(self, name=name)
         self.lower_bound = lower_bound
@@ -771,6 +792,7 @@ class Parameter(DessiaObject):
 
 class ParameterSet(DessiaObject):
     """ Object that can provide utils features around values Dataset. """
+
     def __init__(self, values, name=''):
         self.values = values
 
@@ -784,7 +806,7 @@ class ParameterSet(DessiaObject):
 
     @property
     def means(self):
-        """ Compute means on all parameters defined in Set"""
+        """ Compute means on all parameters defined in Set. """
         means = {k: sum(v) / len(v) for k, v in self.values.items()}
         return means
 
@@ -1256,7 +1278,13 @@ def prettyname(namestr):
 def inspect_arguments(method, merge=False):
     """ Get method arguments and default arguments as sequences while removing forbidden ones (self, cls...)."""
     # Find default value and required arguments of class construction
-    argspecs = inspect.getfullargspec(method)
+    method_full_name = f'{method.__module__}.{method.__qualname__}'
+    if method_full_name in _fullargsspec_cache:
+        argspecs = _fullargsspec_cache[method_full_name]
+    else:
+        argspecs = inspect.getfullargspec(method)
+        _fullargsspec_cache[method_full_name] = argspecs
+
     nargs, ndefault_args = split_argspecs(argspecs)
 
     default_arguments = {}