+++ dessia_common/exports.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-exports for dessia_common
-
+Exports for dessia_common.
 """
 from typing import List, Dict, Any, Sequence, Optional
 
@@ -38,7 +37,7 @@ def is_builtins_list(list_):
 
 
 class ExportFormat:
-    """ Define which method of an object should be called for each Export """
+    """ Define which method of an object should be called for each Export. """
     def __init__(self, selector: Optional[str], extension: str, method_name: str, text: bool,
                  export_name: str = "", args: Dict[str, Any] = None):
         self.selector = selector
@@ -256,6 +255,7 @@ class XLSXWriter:
 
 class MarkdownWriter:
     """ Base class to write markdowns. """
+
     def __init__(self, print_limit: int = 25, table_limit: int = 12):
         self.print_limit = print_limit
         self.table_limit = table_limit