+++ dessia_common/utils/jsonschema.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-JsonSchema generation functions
+JsonSchema generation functions.
 """
 from copy import deepcopy
 import inspect