+++ dessia_common/utils/serialization.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Serialization Tools
-
+Serialization Tools.
 """
 
 import warnings
@@ -11,7 +10,7 @@ import dessia_common.serialization as dcs
 
 def serialize_dict(dict_):
     """
-    Serialize a dict into a dict (values are serialized)
+    Serialize a dict into a dict (values are serialized).
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize_dict(dict_)
@@ -19,7 +18,7 @@ def serialize_dict(dict_):
 
 def serialize_sequence(seq):
     """
-    Serialize a sequence (list or sequence) into a list of dicts
+    Serialize a sequence (list or sequence) into a list of dicts.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize_sequence(seq)
@@ -27,8 +26,7 @@ def serialize_sequence(seq):
 
 def serialize(value):
     """
-    Main function for serialization without pointers
-    Calls recursively itself serialize_sequence and serialize_dict
+    Main function for serialization without pointers. Calls recursively itself serialize_sequence and serialize_dict.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize(value)
@@ -36,7 +34,7 @@ def serialize(value):
 
 def serialize_with_pointers(value, memo=None, path='#'):
     """
-    Main function for serialization with pointers
+    Main function for serialization with pointers.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize_with_pointers(value=value, memo=memo, path=path)
@@ -44,7 +42,7 @@ def serialize_with_pointers(value, memo=None, path='#'):
 
 def serialize_dict_with_pointers(dict_, memo, path):
     """
-    Serialize a dict recursively with jsonpointers using a memo dict at a given path of the top level object
+    Serialize a dict recursively with jsonpointers using a memo dict at a given path of the top level object.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize_dict_with_pointers(dict_=dict_, memo=memo, path=path)
@@ -52,7 +50,7 @@ def serialize_dict_with_pointers(dict_, memo, path):
 
 def serialize_sequence_with_pointers(seq, memo, path):
     """
-    Serialize a sequence (list or tuple) using jsonpointers
+    Serialize a sequence (list or tuple) using jsonpointers.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.serialize_sequence_with_pointers(seq=seq, memo=memo, path=path)
@@ -61,7 +59,7 @@ def serialize_sequence_with_pointers(seq, memo, path):
 def deserialize(serialized_element, sequence_annotation: str = 'List', global_dict=None,
                 pointers_memo=None, path: str = '#'):
     """
-    Main function for deserialization, handle pointers
+    Main function for deserialization, handle pointers.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialize(serialized_element=serialized_element, sequence_annotation=sequence_annotation,
@@ -69,6 +67,7 @@ def deserialize(serialized_element, sequence_annotation: str = 'List', global_di
 
 
 def deserialize_sequence(sequence, annotation=None, global_dict=None, pointers_memo=None, path='#'):
+    """ Deserialization function for sequences (tuple + list). """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialize_sequence(sequence=sequence, annotation=annotation, global_dict=global_dict,
                                     pointers_memo=pointers_memo, path=path)
@@ -77,7 +76,9 @@ def deserialize_sequence(sequence, annotation=None, global_dict=None, pointers_m
 def dict_to_object(dict_, class_=None, force_generic: bool = False,
                    global_dict=None, pointers_memo=None, path='#'):
     """
-    Transform a dict to an object
+    Transform a dict to an object.
+
+    :param global_dict: The global object to find references into, where # is the root.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.dict_to_object(dict_=dict_, class_=class_, force_generic=force_generic, global_dict=global_dict,
@@ -85,13 +86,20 @@ def dict_to_object(dict_, class_=None, force_generic: bool = False,
 
 
 def deserialize_with_type(type_, value):
+    """
+    Deserialization a value enforcing the right type.
+
+    Usefull to deserialize as a tuple and not a list for ex.
+    """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialize_with_type(type_=type_, value=value)
 
 
 def deserialize_with_typing(type_, argument, global_dict=None, pointers_memo=None, path='#'):
     """
-    Deserialize an object with a typing info
+    Deserialize an object with a typing info.
+
+    :param path: The path of the argument in global dict.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialize_with_typing(type_=type_, argument=argument, global_dict=global_dict,
@@ -100,7 +108,9 @@ def deserialize_with_typing(type_, argument, global_dict=None, pointers_memo=Non
 
 def deserialize_argument(type_, argument, global_dict=None, pointers_memo=None, path='#'):
     """
-    Deserialize an argument of a function with the type
+    Deserialize an argument of a function with the type.
+
+    :param path: The path of the argument in global dict.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialize_argument(type_=type_, argument=argument, global_dict=global_dict,
@@ -109,26 +119,37 @@ def deserialize_argument(type_, argument, global_dict=None, pointers_memo=None, 
 
 def find_references(value, path='#'):
     """
-    Traverse recursively the value to find reference (pointers) in it
+    Traverse recursively the value to find reference (pointers) in it.
+
     Calls recursively find_references_sequence and find_references_dict
+
+    :param path: The path of the value in case of a recursive search.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.find_references(value=value, path=path)
 
 
 def find_references_sequence(seq, path):
+    """
+    Compute references for a sequence.
+    """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.find_references_sequence(seq=seq, path=path)
 
 
 def find_references_dict(dict_, path):
+    """
+    Compute references for a dict.
+    """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.find_references_dict(dict_=dict_, path=path)
 
 
 def pointer_graph(value):
     """
-    Create a graph of subattributes of an object with edge representing either:
+    Create a graph of subattributes of an object.
+
+    For this function, edge representing either:
      * the hierarchy of an subattribute to an attribute
      * the pointer link between the 2 elements
     """
@@ -137,13 +158,16 @@ def pointer_graph(value):
 
 
 def update_pointers_data(global_dict, current_dict, pointers_memo):
+    """
+    Update data dict from pointer.
+    """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.update_pointers_data(global_dict=global_dict, current_dict=current_dict, pointers_memo=pointers_memo)
 
 
 def deserialization_order(dict_):
     """
-    Analyse a dict representing an object and give a deserialization order
+    Analyse a dict representing an object and give a deserialization order.
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.deserialization_order(dict_)
@@ -151,9 +175,12 @@ def deserialization_order(dict_):
 
 def dereference_jsonpointers(dict_):
     """
-    Analyse the given dict to:
+    Analyse the given dict.
+
+    It searches for:
     - find jsonpointers
     - deserialize them in the right order to respect pointers graph
+
     :returns: a dict with key the path of the item and the value is the python object
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
@@ -161,26 +188,27 @@ def dereference_jsonpointers(dict_):
 
 
 def pointer_graph_elements(value, path='#'):
+    """ Compute pointer graph. """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.pointer_graph_elements(value=value, path=path)
 
 
 def pointer_graph_elements_sequence(seq, path='#'):
-    """
-    Compute
-    """
+    """ Compute pointer sequence graph. """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.pointer_graph_elements_sequence(seq=seq, path=path)
 
 
 def pointer_graph_elements_dict(dict_, path='#'):
+    """ Compute pointer dict graph. """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")
     return dcs.pointer_graph_elements_dict(dict_=dict_, path=path)
 
 
 def pointers_analysis(obj):
     """
-    Analyse on object to output stats on pointer use in the object
+    Analyse on object to output stats on pointer use in the object.
+
     :returns: a tuple of 2 dicts: one giving the number of pointer use by class
     """
     warnings.warn("Module serialization.py have been moved outside utils. Please use it instead")