+++ dessia_common/workflow/blocks.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Mar 18 18:52:32 2022
-
-@author: masfaraud
+Module to define Blocks for workflows.
 """
 
 import inspect
@@ -31,7 +29,9 @@ from dessia_common.workflow.utils import ToScriptElement
 
 
 def set_inputs_from_function(method, inputs=None):
-    """ Inspect given method argspecs and sets block inputs from it. """
+    """
+    Inspect given method argspecs and sets block inputs from it.
+    """
     if inputs is None:
         inputs = []
     args_specs = inspect.getfullargspec(method)
@@ -55,7 +55,9 @@ def set_inputs_from_function(method, inputs=None):
 
 
 def output_from_function(function, name: str = "result output"):
-    """ Inspect given function argspecs and compute block output from it. """
+    """
+    Inspect given function argspecs and compute block output from it.
+    """
     annotations = get_type_hints(function)
     if 'return' in annotations:
         type_ = type_from_annotation(annotations['return'], function.__module__)
@@ -64,7 +66,9 @@ def output_from_function(function, name: str = "result output"):
 
 
 class BlockError(Exception):
-    """ Specific BlockError Exception. """
+    """
+    Specific BlockError Exception.
+    """
 
 
 class InstantiateModel(Block):
@@ -88,11 +92,15 @@ class InstantiateModel(Block):
         return len(self.model_class.__name__)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         classname = self.model_class.__class__.__name__
         other_classname = other.model_class.__class__.__name__
         return Block.equivalent(self, other) and classname == other_classname
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['model_class'] = full_classname(object_=self.model_class, compute_for='class')
         return dict_
@@ -138,6 +146,7 @@ class InstantiateModel(Block):
 
 class ClassMethod(Block):
     """ Run given classmethod during workflow execution. Handle static method as well. """
+
     def __init__(self, method_type: ClassMethodType[Type], name: str = '', position=None):
         self.method_type = method_type
         inputs = []
@@ -155,6 +164,9 @@ class ClassMethod(Block):
         return len(classname) + 7 * len(self.method_type.name)
 
     def equivalent(self, other: 'ClassMethod'):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         classname = self.method_type.class_.__name__
         other_classname = other.method_type.class_.__name__
         same_class = classname == other_classname
@@ -221,6 +233,7 @@ class ModelMethod(Block):
     :param name: Name of the block.
     :type name: str
     """
+
     def __init__(self, method_type: MethodType[Type], name: str = '', position=None):
         self.method_type = method_type
         inputs = [TypedVariable(type_=method_type.class_, name='model at input')]
@@ -245,6 +258,9 @@ class ModelMethod(Block):
         return len(classname) + 7 * len(self.method_type.name)
 
     def equivalent(self, other: 'ModelMethod'):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         classname = self.method_type.class_.__name__
         other_classname = other.method_type.class_.__name__
         same_model = classname == other_classname
@@ -252,6 +268,7 @@ class ModelMethod(Block):
         return Block.equivalent(self, other) and same_model and same_method
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         classname = full_classname(object_=self.method_type.class_, compute_for='class')
         method_type_dict = {'class_': classname, 'name': self.method_type.name}
@@ -312,6 +329,7 @@ class Sequence(Block):
     :param name: Block name.
     :param position: Position in canvas.
     """
+
     def __init__(self, number_arguments: int, name: str = '', position=None):
         self.number_arguments = number_arguments
         inputs = [Variable(name=f"Sequence element {i}") for i in range(self.number_arguments)]
@@ -322,9 +340,13 @@ class Sequence(Block):
         return self.number_arguments
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.number_arguments == other.number_arguments
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['number_arguments'] = self.number_arguments
         return dict_
@@ -354,9 +376,13 @@ class Concatenate(Block):
         return self.number_arguments
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.number_arguments == other.number_arguments
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['number_arguments'] = self.number_arguments
         return dict_
@@ -407,11 +433,15 @@ class WorkflowBlock(Block):
         return hash(self.workflow)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         if not Block.equivalent(self, other):
             return False
         return self.workflow == other.workflow
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_.update({'workflow': self.workflow.to_dict(use_pointers=use_pointers, memo=memo, path=f'{path}/workflow')})
         return dict_
@@ -491,11 +521,15 @@ class ForEach(Block):
         return wb_hash + self.iter_input_index
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         input_eq = self.iter_input_index == other.iter_input_index
         wb_eq = self.workflow_block.equivalent(other.workflow_block)
         return Block.equivalent(self, other) and wb_eq and input_eq
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         wb_dict = self.workflow_block.to_dict(use_pointers=use_pointers, memo=memo, path=f"{path}/worklow_block")
         dict_.update({'workflow_block': wb_dict, 'iter_input_index': self.iter_input_index})
@@ -541,18 +575,23 @@ class ForEach(Block):
 
 class Unpacker(Block):
     """ DeMUX block. """
+
     def __init__(self, indices: List[int], name: str = '', position=None):
         self.indices = indices
         outputs = [Variable(name=f"output_{i}") for i in indices]
         Block.__init__(self, inputs=[Variable(name="input_sequence")], outputs=outputs, name=name, position=position)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.indices == other.indices
 
     def equivalent_hash(self):
         return len(self.indices)
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['indices'] = self.indices
         return dict_
@@ -606,9 +645,13 @@ class Product(Block):
         return self.number_list
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.number_list == other.number_list
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['number_list'] = self.number_list
         return dict_
@@ -622,7 +665,7 @@ class Product(Block):
 
     def evaluate(self, values, **kwargs):
         """
-        Computes the block: use itertools.product
+        Computes the block: use itertools.product.
         """
         list_product = [values[var] for var in self.inputs]
         output_value = list(itertools.product(*list_product))
@@ -635,6 +678,8 @@ class Product(Block):
 
 class Filter(Block):
     """
+    A Block to filter some data according to some filters.
+
     :param filters: A list of dictionaries, each corresponding to a value to filter.
                     The dictionary should be as follows :
                     *{'attribute' : Name of attribute to filter (str),
@@ -654,6 +699,9 @@ class Filter(Block):
         Block.__init__(self, inputs, outputs, name=name, position=position)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.filters == other.filters
 
     def equivalent_hash(self):
@@ -661,6 +709,7 @@ class Filter(Block):
         return int(sum(hashes) % 10e5)
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         filters_dict = [f.to_dict(use_pointers=use_pointers, memo=memo, path=f"{path}/filters/{i}")
                         for i, f in enumerate(self.filters)]
@@ -762,6 +811,7 @@ class MultiPlot(Display):
     :param name: The name of the block.
     :type name: str
     """
+
     def __init__(self, attributes: List[str], order: int = None, name: str = '', position=None):
         if order is not None:
             warnings.warn("Display Block : order argument is deprecated and will be removed in a future version."
@@ -774,6 +824,9 @@ class MultiPlot(Display):
         self._selector = None
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         same_attributes = self.attributes == other.attributes
         return Block.equivalent(self, other) and same_attributes
 
@@ -781,6 +834,7 @@ class MultiPlot(Display):
         return sum(len(a) for a in self.attributes)
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['attributes'] = self.attributes
         return dict_
@@ -853,8 +907,7 @@ class Markdown(Display):
 
 class PlotData(Display):
     """
-    Generate a DisplayObject that is displayable in PlotData features.
-    Uses the the input object's plot_data method.
+    Generate a DisplayObject that is displayable in PlotData features. Uses the the input object's plot_data method.
 
     :param name: The name of the block.
     :type name: str
@@ -888,9 +941,13 @@ class ModelAttribute(Block):
         return len(self.attribute_name)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.attribute_name == other.attribute_name
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_.update({'attribute_name': self.attribute_name})
         return dict_
@@ -911,6 +968,8 @@ class ModelAttribute(Block):
 
 class SetModelAttribute(Block):
     """
+    Block to set an attribute value in a workflow.
+
     :param attribute_name: The name of the attribute to set.
     :type attribute_name: str
     :param name: The name of the block.
@@ -927,9 +986,13 @@ class SetModelAttribute(Block):
         return 3 + len(self.attribute_name)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.attribute_name == other.attribute_name
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_.update({'attribute_name': self.attribute_name})
         return dict_
@@ -960,9 +1023,13 @@ class Sum(Block):
         return self.number_elements
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         return Block.equivalent(self, other) and self.number_elements == other.number_elements
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_.update({'number_elements': self.number_elements})
         return dict_
@@ -1016,11 +1083,15 @@ class ConcatenateStrings(Block):
         return self.number_elements + hash(self.separator)
 
     def equivalent(self, other):
+        """
+        Returns if the block is equivalent to the other given.
+        """
         same_number = self.number_elements == other.number_elements
         same_separator = self.separator == other.separator
         return Block.equivalent(self, other) and same_number and same_separator
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_.update({'number_elements': self.number_elements, "separator": self.separator})
         return dict_
@@ -1061,6 +1132,7 @@ class Export(Block):
     :param name: Name of the block.
     :type name: str
     """
+
     def __init__(self, method_type: MethodType[Type], text: bool, extension: str,
                  filename: str = "export", name: str = "", position=None):
         self.method_type = method_type
@@ -1079,6 +1151,7 @@ class Export(Block):
         Block.__init__(self, inputs=inputs, outputs=[output], name=name, position=position)
 
     def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         classname = full_classname(object_=self.method_type.class_, compute_for='class')
         method_type_dict = {'class_': classname, 'name': self.method_type.name}
@@ -1136,6 +1209,7 @@ class Archive(Block):
     :param name: Name of the block.
     :type name: str
     """
+
     def __init__(self, number_exports: int = 1, filename: str = "archive", name: str = "", position=None):
         self.number_exports = number_exports
         self.filename = filename
@@ -1146,6 +1220,7 @@ class Archive(Block):
         Block.__init__(self, inputs=inputs, outputs=[Variable(name="zip archive")], name=name, position=position)
 
     def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#'):
+        """Serialize the block with custom logic."""
         dict_ = Block.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
         dict_['number_exports'] = len(self.inputs) - 1   # Filename is also a block input
         dict_["filename"] = self.filename