+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.12.0 (now branch dev)
+
+### Fixed
+- Hot fix for is_valid method of workflow's Block
+
+### CI
+
+- add a check to enforce update of changelog in PR
+
+### Performance
+
+- Conform doc for many parts of dessia_common
+- 100% coverage for clusters module
+- cache inspect.fullargspecs calls
+
+
+### Tests
+
+- Add backend backward import tests in order to warn when imports are changed
+
+### Chores
+- Merge back master to dev
+
 ## 0.11.0
 
 ### Fixed